refactor: move tests related to incidents
